Added backdoor button to scan.js
References #36.
Added functionality to scan.js to check if a server is backdoored.
If a non-backdoored server is able to be backdoored, a button appears next to the server's required hack level.
When pressed, the button enters the connect & backdoor command in the terminal.
Right now the backdoor button is a bright green. Perhaps the color should be changed to contrast more with the other outputs?