Always kill all scripts on ascend
Got rid of the "scripts-to-kill" argument, and decided it just makes sense to kill every other script when running ascend.

Without this, there will always be tricky situations, like this recent one where we would kill gangs/sleeve, but then before we could call them again with "spend" args, autopilot started them back up.

If we're about to ascend, every script will be killed anyway, so we might as well do it immediately and have things be stable and predictable.