Improve casino gains stability
Improves on  Fix #112 by not trusting the "casino ran" flag file early in a run if our money is very low. It could be a false positive if all the money was somehow spent, but worst case it just results in a little extra cleanup and getting kicked out of the casino more than once.