fix an issue if reserve.txt becomes really big (#149)
credits go to cwnintendo...

So, think I found a bug in Stats/Helper, specifically `num.toExponential` not existing.
In bn3.3, near the end and started just pumping up the advertisements, and I've got a revenue of $1.550e+54 / sec. Then the stats.js started throwing a bunch of errors, dug it back down to the helper.js.

Dug some more, and it looks like something put an exponential number into reserve.txt. stats.js reads it in as a string, and so in line ~75 when it calls formatNumberShort, it doesn't work because the string doesn't have a toExponential function.