Fix #311, Fix #304, Fix #325 Properly
Global caches persisting between bitnodes are a permanent problem, especially in helpers where they cannot be reset in the main method (which has been the workaround in other scripts).

The only solution is no more caching. Even though we've fixed the two places that were assuming null meant no TIX API access, there's nothing to prevent a similar misunderstanding from happening in the future unless we get rid of the caching behaviour.

I expect the performance won't be much worse, since the only places that use this function are either run once, already ram-dodging the TIX api to get stock amounts and prices (a more expensive calll than this) or are keeping their own cache of symbols themselves (stockmaster) which are properly cleared in the main method. As such, this potentially-stale cache isn't really moving the needle in terms of performance.