Fix #307 - not cracking before destroyW0r1dD43m0n
It's required to have root access to w0r1d_d43m0n before we can invoke ns.singularity.destroyW0r1dD43m0n

This is typically always the case, because daemon.js is on a tight loop that discovers all new servers as they become available and immediately cracks them, but it a race condition is possible (and a user appears to have encountered it) where we detect the win condition and try to destroy the bitnode before it is cracked. This should fix that.