Fix #415 - double check job requirements
Double checked that these are correct. Couldn't reproduce the issue I thought I saw the other day.

Come to think of it, it's probably be because I was working for NWO on behalf of Silhouette, and at the time I hadn't fixed the bit that copies the +25 stat mult over from NWO to Silhouette once we pick them as the company we're working for.